Add new modular firmware configuration
This commit adds a new modular firmware configuration that will become
the baseline configuration for the Milo V1.5.

It is designed to allow certain settings to be configurable from a
`user-vars.g` file but it should run out-of-the-box on Milo V1.5 built
using the LDO kit and wiring.

If you are using a different mainboard or wiring scheme, then you will
need to modify the pin definitions in your own `user-vars.g` file which
will be loaded automatically on boot.

Signed-off-by: Ben Agricola <ben+git@agrico.la>